requirements.txt
hx711==1.1.2.3
PyYAML==6.0.1
RPi.GPIO==0.7.1
spidev==3.6
